+++ src/components/CustomInput/CustomInput.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { withStyles, FormControl, InputLabel, Input } from "material-ui";
-import { Clear, Check } from "material-ui-icons";
+import { Clear, Check } from "@material-ui/icons";
 import PropTypes from "prop-types";
 import cx from "classnames";