+++ src/components/Header/HeaderLinks.jsx
@@ -11,7 +11,7 @@ import {
   ClickAwayListener,
   Hidden
 } from "material-ui";
-import { Person, Notifications, Dashboard, Search } from "material-ui-icons";
+import { Person, Notifications, Dashboard, Search } from "@material-ui/icons";
 
 import { CustomInput, IconButton as SearchButton } from "components";