+++ src/routes/dashboard.jsx
@@ -14,7 +14,7 @@ import {
   BubbleChart,
   LocationOn,
   Notifications
-} from "material-ui-icons";
+} from "@material-ui/icons";
 
 const dashboardRoutes = [
   {